Dns_BombV2/requirements.txt
dnspython>=2.0.0
python-whois>=0.8.0
